abstract slider logic into a composable
